Modified Service to accept PathParameters and builds the full path accordingly. This will enable the ability to utilise features such as sort, limit, page etc.
Adjustment to some of the result POJOs as there were issues with some of the data types and names (for conversion from the APIs json response)

+++ src/main/java/com/example/openaqclient/service/result/AveragingPeriod.java
@@ -2,15 +2,15 @@ package com.example.openaqclient.service.result;
 
 public class AveragingPeriod {
 
-    private int value;
+    private double value;
     private String unit;
 
-    public AveragingPeriod(int value, String unit) {
+    public AveragingPeriod(double value, String unit) {
         this.value = value;
         this.unit = unit;
     }
 
-    public int getValue() {
+    public double getValue() {
         return value;
     }
